fix: satisfy economy repo-lint for labour readiness (#4237)
Use colonizethis_test imports and scenario-table harness in economy
tests; split types into labour_readiness_types.dart to stay under the
310-line economy source ceiling.

Refs #4237
