+++ packages/colonizethis_economy/lib/colonizethis_economy.dart
@@ -32,6 +32,7 @@ export 'src/economy/trade_cargo_capacity.dart';
 export 'src/economy/trade_interception.dart';
 export 'src/economy/worker_action_cost.dart';
 export 'src/economy/labour_readiness.dart';
+export 'src/economy/labour_readiness_types.dart';
 export 'src/economy/worker_economy.dart';
 export 'src/economy/world_market/bid_type_cap.dart';
 export 'src/economy/world_market/boycott_blocked_commodities.dart';
